+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `Ctrl+Tab` and `Ctrl+Shift+Tab` to select next/previous playlist
 - Make time & volume sliders look ok
 - Small design updates to tracklist header & filter textbox
+- Respect user's locale time
 - Fix dragging tracks to up next/autoplay queue boundary
 - Fix button edge not clickable due to zoom
 - Fix rearranging queue items not working when dragging the cover artwork